refactor structure table in project
