+++ src/components/Table/TableForcast/TableForcast.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../Table.module.css";
+import styles from "./TableForecast.module.css";
 import up from "../../../assets/filter-up.svg";
 import down from "../../../assets/filter-down.svg";
 import downActive from "../../../assets/filter-down-active.svg";
