+++ src/components/Table/TableSales/TableSales.tsx
@@ -4,7 +4,7 @@ import up from "../../assets/filter-up.svg";
 import down from "../../assets/filter-down.svg";
 import downActive from "../../assets/filter-down-active.svg";
 import upActive from "../../assets/filter-up-active.svg";
-import { columnsTable, dataSourceTableSales } from "../../types/types";
+import { columnsTable, dataSourceTableSales } from "../../../types/types";
 
 type TableSales = {
   store: string;
